Refactor into modular UI, data, and config
