+++ config/settings.py
@@ -1,0 +1,18 @@
+import json
+import os
+
+CONFIG_FILE = os.path.expanduser("~/.ybs_control_config.json")
+
+def load_config():
+    try:
+        with open(CONFIG_FILE, "r", encoding="utf-8") as f:
+            return json.load(f)
+    except Exception:
+        return {}
+
+def save_config(config):
+    try:
+        with open(CONFIG_FILE, "w", encoding="utf-8") as f:
+            json.dump(config, f)
+    except Exception:
+        pass
